Do some refactor, add some cpp features.

+++ tuntap_if.cpp
@@ -23,11 +23,11 @@ void tun_init() {
     tun_id = fd;
 }
 
-int tun_read(uint8_t* buf, int len) {
+ssize_t tun_read(uint8_t *buf, size_t len) {
     return read(tun_id, buf, len);
 }
 
-int tun_write(uint8_t* buf, int len) {
+ssize_t tun_write(uint8_t *buf, size_t len) {
     return write(tun_id, buf, len);
 }
